+++ CHANGELOG.md
@@ -1,6 +1,10 @@
+## 6.1.0
+
+* Allow configuring border radius separately for the navigation bar container and its buttons
+
 ## 6.0.0
 
-* BREAKING: Upgrade minimum Flutter version to 3.29 due to using Color.withValues() 
+* BREAKING: Upgrade minimum Flutter version to 3.29 due to using Color.withValues()
 * Update Android tooling
 
 ## 5.1.0